[feat?]: exit the program if errors are found. which is hard.
I can't exit the program the right way, apparently. This commit exits main in a really fun way if there are errors.

it's not graceful.

And comment out "call exit" in apex_routines, since there is no exit here.

This is super hacky and not "correct" (or "good").  But it works!